# Faster blit_with_alpha()