… games/devtest/mods/testnodes/textures/testnodes_alpha_compositing_bottom.png
… games/devtest/mods/testnodes/textures/testnodes_alpha_compositing_top.png
… games/devtest/mods/testnodes/textures/testnodes_fill_positioning_reference.png