@HybridDog 2024-03-13
I have tested the performance with different implementation decisions now by collecting the `blit_with_alpha` inputs and measuring the time it takes to re-execute all the `blit_with_alpha`s.

I tried it 9 times and there were 1687 `blit_with_alpha` inputs each time.
`for f in {1..9}; do bin/minetest --name your_nameaaa --address 127.0.0.1 --go --quiet; done`

#

With the branches the median duration was ca. 0.037 s.
All durations: 0.0353195, 0.0358201, 0.0363832, 0.0364065, 0.0372696, 0.0373504, 0.0381849, 0.0395097, 0.040282
Code: https://github.com/HybridDog/minetest/tree/8cbd75b076bb5d91caed2aaf90121b7753675d13

#

Without the branches the median was ca. 0.082 s.
All durations: 0.0795838, 0.0796296, 0.0812277, 0.08147, 0.0819993, 0.0820157, 0.082015, 0.0822578, 0.089137
Code: https://github.com/HybridDog/minetest/tree/efbce26

#

With the branches, a small reordering, and using SColor, the median was ca. 0.026 s.
The small reordering is the movement of the branch `src_a == 0` before the branch `src_a == 255 || dst_a == 0`; without this the median was ca. 0.035 s.
I also tried u32 integers instead of static_casts to u8 and it was noticeable faster with u8 integers.
The code from this case is used in this Pull Request.
All durations: 0.0247669, 0.0249853, 0.0254656, 0.0255048, 0.0257042, 0.026188, 0.026427, 0.0265187, 0.0272248
Code: https://github.com/HybridDog/minetest/tree/96505ee7f

#

With `blit_with_alpha` from the master branch, the median was ca. 0.384 s.
All durations: 0.361643, 0.368806, 0.373106, 0.380362, 0.384423, 0.386407, 0.393304, 0.394426, 0.401159
Code: https://github.com/HybridDog/minetest/tree/6c7009384

#